backend/requirements.txt
fastapi==0.115.8
uvicorn[standard]==0.34.0
pydantic==2.10.6
python-dotenv==1.0.1
groq==0.18.0
numpy==2.2.3
httpx==0.28.1
firebase-admin==6.6.0
